Fixed file sending problem between TCP Client and TCP Server
Server was waiting for a string from the client but i removed the 'hello
server' from the client since it was not useful. Added mainStation.py
and mainRobot.py to start having some "real-life" test on the table.
Next step: Working on the Linux script to start on embedded computer
boot.